+++ tests/testthat/test-build-reference-index.R
@@ -57,11 +57,7 @@ test_that("warns if missing topics", {
   meta <- list(reference = ref)
   pkg <- as_pkgdown(test_path("assets/reference"), override = meta)
 
-  withr::local_envvar(c(CI = "false"))
-  expect_warning(data_reference_index(pkg), "Topics missing")
-
-  withr::local_envvar(c(CI = "true"))
-  expect_error(data_reference_index(pkg), "Topics missing")
+  expect_snapshot(data_reference_index(pkg), error = TRUE)
 })
 
 test_that("default reference includes all functions", {